Arreglos 1

+++ veterinaria/OhMyDog/modelos/perros/perros.py
@@ -8,7 +8,7 @@ class Perro(models.Model):
     dueño = models.ForeignKey(Cliente, null=False, on_delete=models.CASCADE)
     nombre = models.CharField(max_length=35, blank=False)
     raza = models.CharField(max_length=255, blank=False)
-    peso = models.IntegerField(blank=False)
+    peso = models.DecimalField(blank=False, max_digits=5, decimal_places=2)
     descripcion = models.CharField(max_length=255)
     fecha_de_nacimiento = models.DateField(blank=False)
     sexos = [("M", "Macho"), ("H", "Hembra")]
